pyproject.toml
[tool.poetry]
name = "uptime_monitor"
version = "0.1.0"
description = ""
authors = ["melhin"]

[tool.poetry.dependencies]
python = "^3.7"
marshmallow = "^3.14.0"
kafka-python = "^2.0.2"
psycopg2 = "^2.9.1"
pydantic = "^1.8.2"
apscheduler = "^3.8.1"
httpx = "^0.20.0"
aiokafka = "^0.7.2"
aiopg = "^1.3.3"

[tool.poetry.dev-dependencies]
responses = "^0.15.0"
pytest-asyncio = "^0.16.0"
pytest-httpx = "^0.14.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
